refactor!: move `BlockTrackerPollingController` to `user-operation-controller`
