+++ packages/polling-controller/package.json
@@ -54,7 +54,6 @@
   },
   "dependencies": {
     "@metamask/base-controller": "^9.1.0",
-    "@metamask/network-controller": "^35.0.1",
     "@metamask/utils": "^11.11.0",
     "@types/uuid": "^8.3.0",
     "fast-json-stable-stringify": "^2.1.0",
