+++ packages/user-operation-controller/src/BlockTrackerPollingController.ts
@@ -4,12 +4,12 @@ import type {
   NetworkClient,
 } from '@metamask/network-controller';
 import type { Json } from '@metamask/utils';
-
 import {
   AbstractPollingControllerBaseMixin,
   getKey,
-} from './AbstractPollingController.js';
-import type { Constructor, PollingTokenSetId } from './types.js';
+  Constructor,
+  PollingTokenSetId,
+} from '@metamask/polling-controller';
 
 /**
  * The minimum input required to start polling for a {@link BlockTrackerPollingController}.
